Add CAI Parameters to Image Parameters
Stability should sign image artifacts using the C2PA tool if
CAI Parameters are set in the request. This commit creates a new CAI Parameters
message, which contains the model name and version. This message can be optionally
included in Image Parameters to get C2PA signing of the returned image artifacts.