Change CAI Parameters message to not take model name or version
Model name and version can be extracted from the engine id. Consequently,
we should not make it a parameter in the request. Instead CAI Parameters
allows users to specify the type of parameters which can be passed in. Currently
the only option is an enum representing the type of model metadata
which will be attatched to the image.